pyproject.toml
[project]
name = "twitter-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "pip>=25.3",
    "psycopg2>=2.9.11",
    "pyjwt>=2.10.1",
    "werkzeug>=3.1.4",
]
